+++ Changelog.md
@@ -1,6 +1,8 @@
 ## OpenStudio-HPXML v1.4.0
+
 __New Features__
 - Updates to OpenStudio 3.4.0/EnergyPlus 22.1.
+- **Breaking change**: New required `OccupancyCalculationType` input to specify operational vs asset calculation. If operational, `NumberofResidents` is required.
 - Expanded capabilities for scheduling:
   - Allows modeling detailed HVAC setpoints via a schedule CSV file.
 - New heat pump capabilities: